requirements.txt
flask==3.1.0
flask-socketio==5.5.1
google-generativeai==0.8.5
elevenlabs==2.11.0
SpeechRecognition==3.14.3
audioop-lts==0.2.2
python-dotenv==1.1.0
gunicorn==23.0.0
requests==2.32.3
